Update patch set 1
Patch Set 1:

I don't know if that is something you are interested in but it looks pretty sad if there are no wallpapers listed in ThemePicker. This way the wallpapers are provided through this simple partner app to ThemePicker. The small papers (currently in following commit) are required for the previews, without them the wallpapers are not listed at all.

Patch-set: 1